- adjusted spell pathes & changed them to 16 (saved 3*16bit in the object)

+++ trunk/server/src/common/button.c
@@ -582,7 +582,7 @@ void add_button_link(object *button, mapstruct *map, int connected)
         LOG(llevBug, "BUG: Tried to add button-link without map.\n");
         return;
     }
-    button->path_attuned = connected;  /* peterm:  I need this so I can rebuild 
+    button->path_attuned = (uint16) connected;  /* peterm:  I need this so I can rebuild 
                                        a connected map from a template map. */
     /*  LOG(llevDebug,"adding button %s (%d)\n", button->name, connected);*/
